renderTask is a part of Task class from now on
