+++ src/main.js
@@ -1,8 +1,7 @@
 import BoardComponent from "./components/board.js";
 import FilterComponent from "./components/filter.js";
 import LoadMoreButtonComponent from "./components/more-button.js";
-import TaskEditComponent from "./components/task-edit.js";
-import TaskComponent from "./components/task.js";
+
 import TasksComponent from "./components/tasks.js";
 import SiteMenuComponent from "./components/site-menu.js";
 import SortComponent from "./components/sort.js";
@@ -10,48 +9,49 @@ import NoTasksComponent from "./components/no-tasks.js";
 import {generateTasks} from "./mock/task.js";
 import {generateFilters} from "./mock/filter.js";
 import {render, RenderPosition} from "./utils.js";
+import Task from "./components/task.js";
 
 
 const TASK_COUNT = 22;
 const SHOWING_TASKS_COUNT_ON_START = 8;
 const SHOWING_TASKS_COUNT_BY_BUTTON = 8;
 
-const renderTask = (taskListElement, task) => {
-  const replaceTaskToEdit = () => {
-    taskListElement.replaceChild(taskEditComponent.getElement(), taskComponent.getElement());
-  };
+// const renderTask = (taskListElement, task) => {
+//   const replaceTaskToEdit = () => {
+//     taskListElement.replaceChild(taskEditComponent.getElement(), taskComponent.getElement());
+//   };
 
 
-  const replaceEditToTask = () => {
-    taskListElement.replaceChild(taskComponent.getElement(), taskEditComponent.getElement());
-  };
+//   const replaceEditToTask = () => {
+//     taskListElement.replaceChild(taskComponent.getElement(), taskEditComponent.getElement());
+//   };
 
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
+//   const onEscKeyDown = (evt) => {
+//     const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
 
-    if (isEscKey) {
-      replaceEditToTask();
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
+//     if (isEscKey) {
+//       replaceEditToTask();
+//       document.removeEventListener(`keydown`, onEscKeyDown);
+//     }
+//   };
 
-  const taskComponent = new TaskComponent(task);
-  const editButton = taskComponent.getElement().querySelector(`.card__btn--edit`);
-  editButton.addEventListener(`click`, () => {
-    replaceTaskToEdit();
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
+//   const taskComponent = new TaskComponent(task);
+//   const editButton = taskComponent.getElement().querySelector(`.card__btn--edit`);
+//   editButton.addEventListener(`click`, () => {
+//     replaceTaskToEdit();
+//     document.addEventListener(`keydown`, onEscKeyDown);
+//   });
 
-  const taskEditComponent = new TaskEditComponent(task);
-  const editForm = taskEditComponent.getElement().querySelector(`form`);
-  editForm.addEventListener(`submit`, (evt) => {
-    evt.preventDefault();
-    replaceEditToTask();
-    document.removeEventListener(`keydown`, onEscKeyDown);
-  });
+//   const taskEditComponent = new TaskEditComponent(task);
+//   const editForm = taskEditComponent.getElement().querySelector(`form`);
+//   editForm.addEventListener(`submit`, (evt) => {
+//     evt.preventDefault();
+//     replaceEditToTask();
+//     document.removeEventListener(`keydown`, onEscKeyDown);
+//   });
 
-  render(taskListElement, taskComponent.getElement(), RenderPosition.BEFOREEND);
-};
+//   render(taskListElement, taskComponent.getElement(), RenderPosition.BEFOREEND);
+// };
 
 const renderBoard = (boardComponent, tasks) => {
   const isAllTasksArchived = tasks.every((task) => task.isArchive);
@@ -69,7 +69,7 @@ const renderBoard = (boardComponent, tasks) => {
   let showingTasksCount = SHOWING_TASKS_COUNT_ON_START;
   tasks.slice(0, showingTasksCount)
     .forEach((task) => {
-      renderTask(taskListElement, task);
+      new Task().renderTask(taskListElement, task);
     });
 
   const loadMoreButtonComponent = new LoadMoreButtonComponent();
@@ -80,7 +80,7 @@ const renderBoard = (boardComponent, tasks) => {
     showingTasksCount = showingTasksCount + SHOWING_TASKS_COUNT_BY_BUTTON;
 
     tasks.slice(prevTasksCount, showingTasksCount)
-      .forEach((task) => renderTask(taskListElement, task));
+      .forEach((task) => new Task().renderTask(taskListElement, task));
 
     if (showingTasksCount >= tasks.length) {
       loadMoreButtonComponent.getElement().remove();
